refactor stable structure to use multiple canister classes
